Start to split database_utilities.
