+++ src/estimagic/optimization/optimize.py
@@ -6,11 +6,11 @@ from estimagic.batch_evaluators import process_batch_evaluator
 from estimagic.exceptions import InvalidFunctionError, InvalidKwargsError
 from estimagic.logging.database_utilities import (
     append_row,
-    load_database,
     make_optimization_iteration_table,
     make_optimization_problem_table,
     make_steps_table,
 )
+from estimagic.logging.load_database import load_database
 from estimagic.optimization.check_arguments import check_optimize_kwargs
 from estimagic.optimization.error_penalty import get_error_penalty_function
 from estimagic.optimization.get_algorithm import (
